SearchBox: refactor - extract onHighlight+onSelected as hooks
