+++ src/components/SearchBox/useGetOnHighlight.ts
@@ -1,10 +1,17 @@
 import { AutocompleteHighlightChangeReason } from '@mui/material';
 import { Option } from './types';
 import { getGeocoderSkeleton } from './options/geocoder';
+import { useFeatureContext } from '../utils/FeatureContext';
+import { Feature } from '../../services/types';
 
-export const onHighlightFactory =
-  (setPreview: (feature: unknown) => void) =>
-  (_: never, option: Option, reason: AutocompleteHighlightChangeReason) => {
+export const useGetOnHighlight = () => {
+  const { setPreview } = useFeatureContext();
+
+  return (
+    _: never,
+    option: Option,
+    reason: AutocompleteHighlightChangeReason,
+  ) => {
     if (reason === 'touch') {
       return;
     }
@@ -13,15 +20,16 @@ export const onHighlightFactory =
     }
     if (option.type === 'star' && option.star.center) {
       const { center } = option.star;
-      setPreview({ center });
+      setPreview({ center } as Feature); // TODO fix setPreview to accept only coordinates
       return;
     }
     if (option.type === 'coords') {
       const { center } = option.coords;
-      setPreview({ center });
+      setPreview({ center } as Feature); // TODO fix setPreview to accept only coordinates
       return;
     }
     if (option.type === 'geocoder' && option.geocoder.geometry?.coordinates) {
       setPreview(getGeocoderSkeleton(option));
     }
   };
+};
